bl1: added memory size calculation routines.
This patch changed the structure of getting memory information.
And calculates a timing value that depends on the memory size.

+++ src/devices/memory/ddr3/ddr3_sdram.h
@@ -50,9 +50,9 @@ struct dram_device_info {
 
 	unsigned int column_size;
 	unsigned int row_size;
-	unsigned long bank_size;
-	unsigned short chip_size;
-	unsigned long sdram_size;
+	unsigned long long bank_size;
+	unsigned long long chip_size;
+	unsigned long long sdram_size;
 };
 
 typedef enum {
